Improves invoice payment validation process
Refactors the invoice payment validation process to provide a smoother user experience.

The validation is now triggered by the child component (`RechargeAccountsControl`) upon completion of its internal validation, eliminating the need for a `setTimeout` hack. This ensures that the form is validated only when necessary and prevents premature validation.

Introduces a `isValidating` state to display a spinner while the form is being validated and disable the submit button to prevent multiple submissions. Adds a success message when the form is valid.