Adds approver check to recharge validation
Adds a check to the recharge account validation process to ensure the user is an approver for the given chart string when validating debits.

Uses the UserManager to retrieve the current user and checks if their email is present in the list of approvers for the account.

Also fixes an error message related to invoice approval authorization.